Add ticket creation modal for admin users
A new modal component for ticket creation has been introduced for the admin users. The component includes form controls for setting the user, subject, category, explanation, and an option for adding images to the tickets. Initialized the CreateAdminModalComponent within the tickets admin component. A button to open the modal has been added in the tickets admin HTML. Lastly, the functionality to clear messages when closing the ticket view component has been added for both admin and client view components.