Add username field to Ticket model and UI
A new field 'userName' has been introduced to the Ticket model in order to store and display the user associated with a particular ticket in the admin interface. Code in 'tickets-admin.component.html' was updated to display this new information. Also, an error display line in 'tickets-admin.component.ts' was removed for cleaner error logging.